requirements.txt
black==22.1.0
click==8.0.4
colorama==0.4.4
isort==5.10.1
MouseInfo==0.1.3
mypy-extensions==0.4.3
numpy==1.22.2
opencv-python==4.5.5.62
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
tomli==2.0.1
typing_extensions==4.1.1
